refactor: move schema adapter integration tests
