+++ datafusion/core/tests/integration_tests/schema_adapter/mod.rs
@@ -1,0 +1,1 @@
+mod schema_adapter_integration_tests;
